Revert "Add support for 'podman pod' on FreeBSD"
