--- libpod/pod_internal_freebsd.go
@@ -1,5 +1,0 @@
-package libpod
-
-func (p *Pod) platformRefresh() error {
-	return nil
-}
