--- libpod/runtime_pod_freebsd.go
@@ -1,9 +1,0 @@
-package libpod
-
-import (
-	"github.com/containers/podman/v4/pkg/specgen"
-)
-
-func (r *Runtime) platformMakePod(pod *Pod, p specgen.PodSpecGenerator) error {
-	return nil
-}
