+++ libpod/runtime_pod_unsupported.go
@@ -1,5 +1,5 @@
-//go:build !linux && !freebsd
-// +build !linux,!freebsd
+//go:build !linux
+// +build !linux
 
 package libpod
